+++ operations.h
@@ -2,7 +2,7 @@
 #define OPERATIONS_H
 #include <stdio.h>
 
-unsigned int factorial(unsigned int n);
+double factorial(unsigned int n);
 double combination(unsigned int n, unsigned int r);
 
 #endif
